barebones strcuture: use system for block events, commit command, repo and initialization of repo, use Vector3i for block positions, change plugin class name

+++ src/main/java/org/mattseq/plugin/BlockChange.java
@@ -1,13 +1,13 @@
 package org.mattseq.plugin;
 
-import com.hypixel.hytale.math.vector.Vector3d;
+import com.hypixel.hytale.math.vector.Vector3i;
 
 public class BlockChange {
-    private Vector3d position;
+    private Vector3i position;
     private String newBlockId;
     private final long timestamp;
 
-    public BlockChange(Vector3d position, String newBlockId) {
+    public BlockChange(Vector3i position, String newBlockId) {
         this.position = position;
         this.newBlockId = newBlockId;
         this.timestamp = System.currentTimeMillis();
